requirements.txt
# =============================================================================
# Multimodal Pothole Detection System - Dependencies
# =============================================================================
# Install: pip install -r requirements.txt
# =============================================================================

# Core ML & Deep Learning
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0  # YOLOv8

# Scientific Computing
numpy>=1.21.0
scipy>=1.9.0
pandas>=1.4.0

# Machine Learning
scikit-learn>=1.0.0

# Image Processing
opencv-python>=4.7.0
Pillow>=9.0.0

# Configuration & Utilities
PyYAML>=6.0
python-dotenv>=1.0.0

# Logging & Monitoring
tqdm>=4.64.0

# Optional: GPU Support (uncomment if using CUDA)
# --extra-index-url https://download.pytorch.org/whl/cu118
# torch>=2.0.0+cu118

# Development (optional)
# pytest>=7.0.0
# pytest-cov>=4.0.0
# black>=23.0.0
# flake8>=6.0.0
